+++ source/expression/Expression.ts
@@ -1,4 +1,0 @@
-export abstract class Expression extends String {
-	abstract type: string
-	abstract expression: RegExp
-}